// src/app/shared/interfaces/model.interface.ts
export interface Product {
  id:          number;
  title:       string;
  price:       number;
  description: string;
  category:    string;
  image:       string;
  rating:      Rating;
}

export interface Rating {
  rate:  number;
  count: number;
}
export interface CartItem {
  product:  Product;
  quantity: number;
}